# docs/file-service_openapi.yaml
openapi: 3.0.0
info:
  title: File Service - OSDU
  description: |
    # Key Concepts

    This service allows users to manage files on the  Data Platform. File Management includes uploads, downloads and creation of metadata record for the file.
  version: 1.0.0
servers:
  - url: VALID_FILE_SERVICE_BASE_URL
paths:
...
  /v2/files/{Id}/DownloadURL:
    get:
      tags:
      - File Service
      summary: Gets a URL to download the file
      description: |-
        Gets a URL for downloading the file associated with the unique `id`. URL is only valid for 1 hour.

        **Required roles**: 'users.datalake.viewers' or 'users.datalake.editors' or 'users.datalake.admins' or 'users.datalake.ops'.
      operationId: Gets URL to download the file associated with the given id.
      parameters:
      - name: data-partition-id
        in: header
        description: Specifies the data partition to use. This should either be the
          partition name or crm account ID associated with the partition.
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: opendes
      - name: Id
        in: path
        description: File Metadata record Id.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        200:
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResponse'
        400:
          description: Bad user input. Mandatory fields missing or unacceptable value
            passed to API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        401:
          description: Invalid/Expired Credential
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        404:
          description: Record Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationError'
      security:
        - bearer: []
        - appkey: []
...
components:
  schemas:
...
    DownloadResponse:
      title: Download URL response
      properties:
        SignedURL:
          type: string
...
    ApplicationError:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorModel'
      description: An error that occurs during normal application logic.
    ErrorModel:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
        code:
          type: integer
          format: int32
        message:
          type: string
    Error:
      type: object
      properties:
        message:
          type: string
        reason:
          type: string
        domain:
          type: string
  securitySchemes:
    bearer:
      type: apiKey
      name: Authorization
      in: header
    appkey:
      type: apiKey
      name: appkey
      in: header